pyproject.toml
[project]
name = "warehouse_robot"
version = "0.1.0"
description = "Warehouse robot system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest",
    "pytest-asyncio",
    "paho-mqtt"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]